Exploratory Data Analysis
Revised data wrangling notebook. We're now using only the customer profile dataframe and the transcript dataframe. The offers dataframe complicated matters a little too much. It was much easier to go into EDA with a dataframe in which all the observations represented transactions. Before, some of the rows represented transactions and some represented just an offer received, viewed or completed.